requirements.txt
google-generativeai
python-dotenv
termcolor
googlesearch-python

